Fix M-less polyZ geometries
The M coordinates are optional.  There is no data on the geometry itself
that this is the case; instead, you know this because you run out of
data on the frame reported in the value's header.

This threads that information through the shp parser, keeping it updated
whenever a chunk of data is read out of the stream.

It also adds a "dbg" macro that can be used to print out information
about progress through the parser.  I've left it in because this isn't
the first time I've wanted to annotate every call with some information.